Move kernel to correct folder
